Modified program to link current database schema with application

+++ fireM2/databaseScripts/table_creation.sql
@@ -2,23 +2,23 @@
 
 #table to store mission information
 create table mission
-	(missionID bigint auto_increment,
+	(missionID int auto_increment,
 	missionName varchar(80) not null,
     primary key (missionID));
 
 #table to store general event information
 create table mmEvent
-	(eventID bigint auto_increment,
+	(eventID int auto_increment,
     eventName varchar(80) not null,
-    latitude decimal(10, 6) not null,
-    longitude decimal(10, 6) not null,
-	missionID bigint,
+    latitude float(10, 6) not null,
+    longitude float(10, 6) not null,
+	missionID int,
     primary key (eventID),
     foreign key (missionID) references mission (missionID));
 
 #table to store event state changes
 create table eventState
-	(eventID bigint not null,
+	(eventID int not null,
     updateTime datetime not null,
     state enum('reported', 'assigned', 'in progress', 'on hold', 'completed') not null,
     primary key (eventID, updateTime),
@@ -26,8 +26,8 @@ create table eventState
 
 #table to store written notes about events
 create table eventNote
-	(noteID bigint auto_increment,
-    eventID bigint not null,
+	(noteID int auto_increment,
+    eventID int not null,
     createTime datetime not null,
 	description varchar(120) not null,
     primary key (noteID),
@@ -35,7 +35,7 @@ create table eventNote
     
 #table to store user account information
 create table userAccount
-	(accountID bigint auto_increment,
+	(accountID int auto_increment,
     firstName varchar(80) not null,
     lastName varchar(80),
 	email varchar(80) not null,
